Added configuration properties: use media FPS, default FPS, keep record file, keep temporary files.

+++ src/sune/app/mediadownloader/drm/phase/PostProcessPhase.java
@@ -15,6 +15,7 @@ import sune.app.mediadown.util.Utils;
 import sune.app.mediadownloader.drm.DRMContext;
 import sune.app.mediadownloader.drm.DRMLog;
 import sune.app.mediadownloader.drm.event.PostProcessEvent;
+import sune.app.mediadownloader.drm.integration.DRMPluginConfiguration;
 import sune.app.mediadownloader.drm.process.SimpleAudioProcessor;
 import sune.app.mediadownloader.drm.process.SimpleVideoProcessor;
 import sune.app.mediadownloader.drm.tracker.PostProcessTracker;
@@ -117,9 +118,16 @@ public class PostProcessPhase implements PipelineTask<PostProcessPhaseResult> {
 			DRMProcessUtils.throwIfFailedFFMpeg(process.waitFor());
 		}
 		
-		filesManager.delete(videoOutputPath);
-		filesManager.delete(audioOutputPath);
-		filesManager.delete(outputRecord);
+		DRMPluginConfiguration configuration = DRMPluginConfiguration.instance();
+		
+		if(!configuration.processKeepTemporaryFiles()) {
+			filesManager.delete(videoOutputPath);
+			filesManager.delete(audioOutputPath);
+		}
+		
+		if(!configuration.processKeepRecordFile()) {
+			filesManager.delete(outputRecord);
+		}
 		
 		filesManager.deleteAll();
 	}
